pushing latest changes (debugging seg fault during aggregation)

+++ server/host/host.h
@@ -11,6 +11,6 @@ int host_modelaggregator(unsigned char*** encrypted_accumulator,
         size_t old_params_length,
         unsigned char*** encrypted_new_params_ptr,
         size_t* new_params_length,
-        size_t* contributions);
+        float* contributions);
 
 #endif
